refactor(sync): extract order write payload shaping into Order_Write_Payload
Write_Controller mixed the generic write protocol (envelope/replay/CAS/
checkpoint) with the order-specific shaping of the body forwarded to the
stock wc/v3 orders controller. Move the shaping half verbatim into a new
final Sync\Order_Write_Payload with exactly two public methods:

- for_create()  -> sanitize_order_wc_payload
- for_update()  -> reconcile_order_item_ids, remove_omitted_order_items,
                   reconcile_order_coupon_lines, sanitize_order_wc_payload,
                   drop_unchanged_variation_line_identity (order preserved;
                   the load-bearing 'runs last' comment moves with it)

All six moved methods (plus the misc-sku sentinel helper and its const)
were private with no callers outside Write_Controller, in this repo or in
woocommerce-pos-pro.

for_update() now loads the order ONCE and hands it to the four steps that
each re-loaded it. Each step keeps its own falsy-order guard, so the
aggregate result for an unresolvable id is unchanged (sanitize only).

Write_Controller keeps the protocol half and delegates through a lazily
constructed Order_Write_Payload.
